pyproject.toml
[tool.poetry]
name = "pl-liblab-lotr-sdk"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pl_liblab_lotr_sdk"}]

[tool.poetry.dependencies]
python = "^3.11"
inflection = "^0.5.1"
python-dotenv = "^1.0.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
